- applied review comments - memory tracker x64 fix
