+++ src/unit_tests/engine/ut_resource_managers.cpp
@@ -410,7 +410,6 @@ namespace
 	}
 }
 
-TODO("string tokenizer");
 REGISTER_TEST("unit_tests/engine/material_manager", UT_material_manager, "unit_tests/resource_managers/cisla.tga 262188");
 REGISTER_TEST("unit_tests/engine/material_manager", UT_material_manager, "unit_tests/resource_managers/trava3.dds 2796344");
 REGISTER_TEST("unit_tests/engine/animation_manager", UT_animation_manager, "unit_tests/resource_managers/blender.ani 3424");
